Change order of logic
Changed the way the `if` statement that looks at the package.json and the options passed in by the start script so that it gives precedence to the options set in `scripts/start`. 

Also added in the option to override the name, seeing as specifying the name in the `package.json` is severely limited by the need to have no spaces or capitalized letters.